Set libpq_fetch to restrict to directories too
The way this was done was to run the query once per interesting directories.  I was slightly
concerned about performance overhead but on the whole I think since we issue one query per file
anyway, that an extra 5-10 queries are really nothing especially since they don't significnatly
increase the work the db needs to do.

Change in behavior is not yet documented.  This will be done before I decide how I want to
build this up for submission for commitfest.