Restrict pg_rewind to whitelisted directories.
This is intended to be a minimum working version and in fact builds and passes tests.
Note that tests for extra files have been changed to reflect new behavior and additional
debugging informnation added in to output in case of failure.

The patch iterates through a series of set directories to synchronize them only.  This improves
predictability of the complete state of the system after a rewind.

One important outstanding question here is whether we need to ensure the possibility of backing
up other files if they exist via an --include-path command line switch (this would not be a glob).
In the thread discussing this patch, Michael Paquier has expressed concern about configuration
files created by extensions or other components not being copied.  I could add such a switch but
the patch is long enough, and it is unclear enough to the extent this is needed at present, so
I am leaving it at the reviewer's discretion whether I should add this here or submit a second
patch later to add the ability to add additional paths to the filemap.